Fix id used to find rate constant conditions (#150)
Fixes a bug that prevented rate constant initial conditions from being
displayed properly. The original code was using the index of the element
in the array of rate constant conditions, rather than the id field in
the array element object. These are the same until an array element is
deleted, but afterwards they differ